+++ packages/nft-checkout/src/operations/evm/helpers/estimator/uniswap-v3.ts
@@ -1,4 +1,4 @@
-import { BN } from '@distributedlab/utils'
+import { BN } from '@distributedlab/tools'
 import { IProvider } from '@rarimo/provider'
 import {
   Currency,
@@ -28,24 +28,9 @@ import { computeRealizedPriceImpact } from './uniswap-impact'
 
 const V3_SWAP_DEFAULT_SLIPPAGE = new Percent(250, 10_000)
 
-const getPrice = (
-  from: Token,
-  amount: CurrencyAmount<Currency> | undefined,
-) => {
-  if (!amount || JSBI.equal(amount?.quotient, JSBI.BigInt(0))) {
-    return Price.fromRaw('0', from.decimals, from.symbol)
-  }
-
-  return Price.fromFraction(
-    amount.numerator.toString(),
-    from.decimals,
-    from.symbol,
-  )
-}
-
-const getRoutePath = (route: RouteWithValidQuote[]) => {
+const getRoutePath = (route: RouteWithValidQuote[], isNative: boolean) => {
   return route.reduce((path, r) => {
-    const p = encodeRouteToPath(r.route as Route<Currency, Currency>, true)
+    const p = encodeRouteToPath(r.route as Route<Currency, Currency>, !isNative)
     path += path ? p.replace('0x', '') : p
 
     return path
@@ -53,7 +38,7 @@ const getRoutePath = (route: RouteWithValidQuote[]) => {
 }
 
 const getSwapCurrencyAmount = (token: UNIToken, price: Price) => {
-  return CurrencyAmount.fromRawAmount(token, JSBI.BigInt(getSwapAmount(price)))
+  return CurrencyAmount.fromRawAmount(token, getSwapAmount(price))
 }
 
 const getSlippage = (slippage?: number): Percent => {
@@ -120,11 +105,16 @@ export const estimateUniswapV3 = async (
     from: _from,
     to: _to,
     impact: trade ? computeRealizedPriceImpact(trade) : undefined,
-    price: getPrice(from, amount),
-    path: getRoutePath(route.route),
-    gasPriceInUSD: new BN(estimatedGasUsedUSD.numerator.toString())
-      .fromFraction(estimatedGasUsedUSD.currency.decimals)
-      .toString(),
-    gasPrice: new BN(gasPriceWei.toString()).fromWei().toString(),
+    price: Price.fromBigInt(
+      amount.numerator.toString() ?? '0',
+      _from.decimals,
+      _from.symbol,
+    ),
+    path: getRoutePath(route.route, _from.isNative),
+    gasPriceInUSD: BN.fromBigInt(
+      estimatedGasUsedUSD.numerator.toString(),
+      estimatedGasUsedUSD.currency.decimals,
+    ).toString(),
+    gasPrice: BN.fromBigInt(gasPriceWei.toString(), BN.WEI_DECIMALS).toString(),
   }
 }